+++ common/src/main/java/dev/xylonity/tooltipoverhaul/client/style/TooltipStyle.java
@@ -2,8 +2,6 @@ package dev.xylonity.tooltipoverhaul.client.style;
 
 import dev.xylonity.tooltipoverhaul.client.layer.LayerDepth;
 import dev.xylonity.tooltipoverhaul.client.layer.bridge.*;
-import dev.xylonity.tooltipoverhaul.client.style.background.icon.DefaultBackgroundPreview;
-import dev.xylonity.tooltipoverhaul.client.style.background.icon.VoidBackgroundPreview;
 import dev.xylonity.tooltipoverhaul.client.style.renderer.DefaultArmorStand;
 import dev.xylonity.tooltipoverhaul.client.style.renderer.DefaultIcon;
 import dev.xylonity.tooltipoverhaul.client.style.renderer.DefaultRotatingItem;
@@ -31,11 +29,11 @@ public class TooltipStyle {
     private final ITooltipArmorStand armorStand;
     private final ITooltipRotatingItem rotatingItem;
 
-    public TooltipStyle(ITooltipPanel panel, ITooltipFrame innerFrame, ITooltipEffect effect, ITooltipDividerLine dividerLine) {
+    public TooltipStyle(ITooltipPanel panel, ITooltipFrame innerFrame, ITooltipEffect effect, ITooltipDividerLine dividerLine, ITooltipPreviewBackground iconBackground) {
         this.effect = effect;
         this.panel = panel;
         this.frame = innerFrame;
-        this.previewBackground = new DefaultBackgroundPreview();
+        this.previewBackground = iconBackground;
         this.icon = new DefaultIcon();
         this.text = new DefaultText();
         this.dividerLine = dividerLine;